Hidden_Number.java
Given an array of n integers, The task is to find another integer x .such that, if all the elements of the array are subtracted individually from the number x, then the sum of all the differences should add to 0. If any such integer exists, print the value of x, else print -1.

The idea is to calculate the total sum of the array and divide it by the size of the array. If the sum of the array is perfectly divisible by its size then the quotient obtained from this division operation will be the required hidden number


Sample Input 0

4
2 4 5 7
Sample Output 0

-1